tquic: fix 16 audit findings across security, memory safety, and protocol compliance
Security fixes:
- CF-003: client cert verification now uses correct EKU (client vs server)
- CF-007: OCSP hard-fail mode rejects unverified signatures (fail closed)
- CF-052: retry token address comparison uses crypto_memneq (constant-time)
- CF-039: netfilter hooks use pernet_operations for all namespaces
- CF-077: QUIC-over-TCP returns error instead of init_net fallback

Memory safety fixes:
- CF-130: tquic_migrate_auto() takes path ref inside lock to prevent UAF
- CF-061: server accept error path cleans up CIDs and zero_rtt_buffer
- CF-362: CID sequence number increment moved inside lock

Protocol compliance:
- C-5: ACK delay uses negotiated exponent per RFC 9000 Section 19.3
- H-1: bytes_acked uses actual path MTU instead of hardcoded 1200
- H-5: version greasing added to VN packets per RFC 9000 Section 6.3
- H-7: QUIC v2 retry integrity keys added per RFC 9369 Section 5.8

Code quality:
- CF-150: RSA-PSS hash parsed from AlgorithmIdentifier parameters
- CF-157: u32 casts on byte shifts in 4 locations to prevent signed overflow

+++ net/tquic/crypto/handshake.c
@@ -2740,13 +2740,17 @@ int tquic_hs_process_new_session_ticket(struct tquic_handshake *hs,
 	/* Ticket lifetime */
 	if (p + 4 > end)
 		return -EINVAL;
-	lifetime = (p[0] << 24) | (p[1] << 16) | (p[2] << 8) | p[3];
+	/* CF-157: cast to u32 before shift to avoid signed overflow */
+	lifetime = ((u32)p[0] << 24) | ((u32)p[1] << 16) |
+		   ((u32)p[2] << 8) | (u32)p[3];
 	p += 4;
 
 	/* Ticket age add */
 	if (p + 4 > end)
 		return -EINVAL;
-	age_add = (p[0] << 24) | (p[1] << 16) | (p[2] << 8) | p[3];
+	/* CF-157: cast to u32 before shift to avoid signed overflow */
+	age_add = ((u32)p[0] << 24) | ((u32)p[1] << 16) |
+		  ((u32)p[2] << 8) | (u32)p[3];
 	p += 4;
 
 	/* Ticket nonce */
